Ya realiza reservas

+++ HotelSol2/ReservaHabitacion.cs
@@ -78,7 +78,11 @@ namespace HotelSol2
                 return;
             }
 
+
+
+
             int id_hab = Convert.ToInt32(DGVHabi.CurrentRow.Cells["id_hab"].Value);
+            //int id_hab = ((Habitacion)listahabi[DGVHabi.CurrentRow.Index]).id_hab;
             int id_cliente = Convert.ToInt32(DVGCliente.CurrentRow.Cells["id_cliente"].Value);
             int id_user = Convert.ToInt32(DGVUsers.CurrentRow.Cells["id_user"].Value);
 
@@ -90,7 +94,7 @@ namespace HotelSol2
                 Tipo_pago = CMBTipoPago.SelectedItem.ToString(),
                 Fecha_ent = DTPIngreso.Value,
                 Fecha_sal = DTPSalida.Value,
-                Total_pago = decimal.Parse(TxtTotalPagar.Text)
+                Total_pago = float.Parse(TxtTotalPagar.Text)
             };
             if (mBD.Conectar())
             {
